+++ applications/diem-core/src/server/routes/executors/spark/spark.job.handlers/handle.spark.common.ts
@@ -51,7 +51,10 @@ export const jdbc_config: (type: string) => any = (type: string) => {
  * @returns {(Promise<IConnModel | Error>)}
  * @error {message, connection, trace[]}
  */
-export const getConnection: (org: string, id: string) => Promise<IConnSchema> = async (org: string, id: string): Promise<IConnSchema> => {
+export const getConnection: (org: string, id: string) => Promise<IConnSchema> = async (
+    org: string,
+    id: string
+): Promise<IConnSchema> => {
     const doc: IConnSchema | null = await ConnModel.findOne({ 'project.org': org, alias: id }).lean().exec();
 
     if (doc) {
