+++ applications/diem-core/src/server/routes/executors/spark/spark.job.handlers/handle.spark.custom.ts
@@ -44,7 +44,7 @@ export const handleWithCustom: (doc: IJobSchema, code: string) => Promise<string
                 Array.isArray(doc.job.params.files.loadfiles) &&
                 doc.job.params.files.loadfiles.length > 0
             ) {
-                code = await handleFiles(code, doc.job.params.files.loadfiles);
+                code = await handleFiles(code, doc.job.params.files.loadfiles as { name: string; value: string }[]);
             }
         }
